requirements.txt
python-telegram-bot==21.6
selenium==4.21.0
openpyxl==3.1.3
python-dotenv==1.0.1
chromedriver-autoinstaller==0.6.4
cryptography>=42.0.0
